Navigation on pages reliably works
git-svn-id: http://code.elgg.org/plugins/plugins@4418 36083f99-b078-4883-b0ff-0f9b5a30f544

+++ pages/views/default/pages/sidebar/sidebarthis.php
@@ -13,7 +13,7 @@
 	$entity = $vars['entity'];
 	$parent = $vars['entity']->parent_guid;
 	
-	$currentpage = get_input('currentpage');
+	$currentpage = (int) get_input('currentpage');
 	if ($currentpage) {
 		
 		$path = pages_get_path($currentpage);
